some structural changes, also copiloted readme
