+++ Makefile
@@ -4,9 +4,12 @@ CC    = @cc
 
 FLAGS = -Wall -Wextra -Werror -g -fsanitize=address -fsanitize-address-use-after-scope
 
+INC = -I./includes
+
 SRC = src/core/main.c \
 	src/core/inputs.c \
 	src/parsing/scanner.c \
+	src/data/initialization.c \
 	src/helpers/debug.c
 
 OBJ   = $(patsubst src/%.c, obj/%.o, $(SRC))
@@ -23,7 +26,7 @@ all: obj $(NAME)
 
 $(NAME): $(OBJ)
 	@echo "$(CBOLD)$(YELLOW)    Compiling $(NAME)   $(RESET)"
-	$(CC) $(FLAGS) -lreadline -o $@ $^
+	$(CC) $(FLAGS) $(INC) -lreadline -o $@ $^
 	@echo "$(CBOLD)$(GREEN)      $(NAME) ready!    $(RESET)"
 
 obj:
