+++ README.md
@@ -1,2 +1,36 @@
 # 42Minishell
 Creating a simple shell in C.
+
+## Description
+The objective of this project is to create a simple shell in C using only the following functions:
+- malloc, free
+- access
+- open, close, read, write
+- opendir, readdir, closedir
+- getcwd, chdir
+- stat, lstat, fstat
+- fork, execve
+- wait, waitpid, wait3, wait4
+- signal, kill
+- exit
+
+## Usage
+```
+$> make
+$> ./minishell
+```
+
+## Features
+- Prompt display
+- Command execution
+- Environment variables
+- Builtins: cd, echo, env, setenv, unsetenv, exit
+- Signals: ctrl-C, ctrl-D, ctrl-\
+- Redirections: <, >, <<, >>
+
+## Authors
+- [**Rajh Phuyal**]() - [**Github**]("https://github.com/rajh-phuyal")
+- [**Joao Mirinda**]() - [**Github**]("https://github.com/joaodibba")
+
+## Acknowledgments
+- [**42**](https://www.42.fr/)
