+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: rajphuyal <rajphuyal@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/17 16:11:36 by rajphuyal         #+#    #+#             */
-/*   Updated: 2023/09/30 21:27:09 by rajphuyal        ###   ########.fr       */
+/*   Updated: 2023/09/30 21:40:33 by rajphuyal        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,9 @@ typedef struct s_minishell
 }	t_minishell;
 
 t_minishell	*minishell(void);
+
+// minishell
+int			init_minishell(void);
 
 // parsing gate
 void		lexer(char *input);
