+++ src/parsing/scanner.c
@@ -6,11 +6,11 @@
 /*   By: rajphuyal <rajphuyal@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 23:59:03 by rajphuyal         #+#    #+#             */
-/*   Updated: 2023/09/28 00:02:56 by rajphuyal        ###   ########.fr       */
+/*   Updated: 2023/09/30 21:42:25 by rajphuyal        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/minishell.h"
+#include "minishell.h"
 
 void	lexer(char *input)
 {
